Added:
Picasa - Picasa Photo Organizer will help you organize and edit your digital photos, then create online albums to share with everyone

netdata - Netdata is distributed, real-time, performance and health monitoring for systems and applications. It is a highly optimized monitoring agent you install on all your systems and containers

Go - Go is an open source programming language that makes it easy to build simple, reliable, and efficient software

Web-Maker - Web-Maker is an offline playground for your web experiments

Manuskript - Manuskript is an open-source tool for writers

Polar Bookshelf - Polar Bookshelf is an incremental reading and personal knowledge repository for PDF and the web created using the Electron framework and PDF.js

KeePass - KeePass is a free open source password manager, which helps you to manage your passwords in a secure way

OpenStego - OpenStego is a steganography application

Chrono - Chrono is a time tracking tool written in Go. It is fast and simple to use

twtxt - twtxt is a decentralised, minimalist microblogging service for hackers